Fix IndirectX and don't allow ints to be passed (since they won't wrap).

+++ DaNES.Emulation.Tests/CpuTests/LdaTests.cs
@@ -52,7 +52,7 @@ namespace DanTup.DaNES.Emulation.Tests.CpuTests
 		[InlineData(255, 255, 255, false, true)]
 		public void Lda_Zero_Page_Offset_X(byte value_to_load, byte ram_location, byte offset, bool expectZero, bool expectNegative)
 		{
-			var actual_ram_location = (ram_location + offset) % 256; // Zero page operations wrap around and remain on zero page.
+			var actual_ram_location = (ushort)((ram_location + offset) % 256); // Zero page operations wrap around and remain on zero page.
 			nes.Cpu.Ram.Write(actual_ram_location, value_to_load);
 			nes.Cpu.XRegister = offset;
 			nes.LoadProgram(0xB5, ram_location);
